feat: Move start method to runtimes module
