+++ ichub-backend/runtimes/fastapi/__init__.py
@@ -23,3 +23,5 @@
 # Package-level variables
 __author__ = 'Eclipse Tractus-X Contributors'
 __license__ = "Apache License, Version 2.0"
+
+from .main import start
